code/langgraph-05-multi-agent/requirements.txt
langgraph==1.0.8
langchain-groq==1.1.2
ipython==9.11.0
streamlit==1.54.0
langchain==1.2.12
finnhub-python